% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram_mixing.R
\name{plot_dendrogram_mixing}
\alias{plot_dendrogram_mixing}
\title{Plot dendrogram of mixing measures}
\usage{
plot_dendrogram_mixing(
  dmm,
  dim = 1,
  point_size = 3,
  line_width = 0.5,
  palette = NULL,
  show_legend = FALSE,
  main = NULL
)
}
\arguments{
\item{dmm}{Output from \code{dendrogram_mixing}.}

\item{dim}{Dimension to plot.}

\item{point_size}{Point size.}

\item{line_width}{Line width.}

\item{palette}{Vector of colors.}

\item{show_legend}{Show colors associated with each step in the dendrogram.}

\item{main}{Title name}
}
\value{
A base R plot.
}
\description{
Plot dendrogram of mixing measures
}
