% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram_mixing.R
\name{plot_dendrogram_mixing}
\alias{plot_dendrogram_mixing}
\title{Plot dendrogram of mixing measures}
\usage{
plot_dendrogram_mixing(dmm, dim = 1, atom_size = 1, lwd = 1)
}
\arguments{
\item{dmm}{Output from \code{dendrogram_mixing}.}

\item{dim}{Dimension to plot.}

\item{atom_size}{Point size.}

\item{lwd}{Line width.}
}
\value{
A base R plot.
}
\description{
Plot dendrogram of mixing measures
}
